# Improve performance of some line code