--- a/src/jeznes.c
+++ b/src/jeznes.c
@@ -951,10 +951,10 @@ void draw_pause_sprites(void) {
 }
 
 unsigned char update_line(unsigned char line_index) {
-  temp_byte_6 = get_temp_ptr(struct Line)->flags;
+  set_flags_byte(get_temp_ptr(struct Line)->flags);
 
   // Do nothing if the line is not started.
-  if (!get_line_is_started_flag_from_byte(temp_byte_6)) {
+  if (!get_line_is_started_flag_from_byte(get_flags_byte())) {
     return FALSE;
   }
 
@@ -966,24 +966,21 @@ unsigned char update_line(unsigned char line_index) {
   // When we've reached completion of the tile, move the front of both
   // line segments forward in their directions.
   if (get_temp_ptr(struct Line)->current_block_completion == 8) {
-    set_line_orientation(get_line_orientation_flag_from_byte(temp_byte_6));
+    set_line_orientation(get_line_orientation_flag_from_byte(get_flags_byte()));
     set_tile_index_delta(compute_tile_index_delta(get_line_orientation()));
-    set_negative_line_segment_origin(get_temp_ptr(struct Line)->origin);
 
     // Try and move the front of the negative-direction line segment forward
     // and complete the line segment if it's reached a wall.
-    if (!get_line_is_negative_complete_flag_from_byte(temp_byte_6)) {
+    if (!get_line_is_negative_complete_flag_from_byte(get_flags_byte())) {
       // Before moving the line segment forward, update the metadata for the
       // tile we're moving from.
-      set_current_playfield_index(get_negative_line_segment_origin() -
-                                  get_temp_ptr(struct Line)->tile_step_count *
-                                      get_tile_index_delta());
+      set_current_playfield_index(playfield_tile_from_pixel_coords(get_temp_ptr(struct Line)->negative_sprite_x, get_temp_ptr(struct Line)->negative_sprite_y));
 
       // While it was the front of the line segment, current playfield tile was
       // being drawn as a sprite. Now that it's complete, update the playfield
       // and bg tile. Note: The origin tile already has the playfield flags (and
-      // bg tile) set. We can ignore that one (when tile_step_count == 0).
-      if (get_temp_ptr(struct Line)->tile_step_count != 0) {
+      // bg tile) set. We can ignore that one.
+      if (!get_line_is_first_step_flag_from_byte(get_flags_byte())) {
         set_playfield_tile(
             get_current_playfield_index(),
             get_playfield_tile_type_line(get_line_orientation(), line_index,
@@ -1013,7 +1010,7 @@ unsigned char update_line(unsigned char line_index) {
 
           // Stop when we reach the origin.
           if (get_current_playfield_index() ==
-              get_negative_line_segment_origin()) {
+              get_temp_ptr(struct Line)->origin) {
             break;
           }
         }
@@ -1023,32 +1020,32 @@ unsigned char update_line(unsigned char line_index) {
             get_temp_ptr(struct Line)->flags);
 
         // When line segments in both directions are complete, the line is done.
-        if (get_line_is_positive_complete_flag_from_byte(temp_byte_6)) {
+        if (get_line_is_positive_complete_flag_from_byte(get_flags_byte())) {
           unset_line_is_started_flag_in_byte(get_temp_ptr(struct Line)->flags);
         }
 
         set_was_line_segment_completed(TRUE);
+      } else {
+        // If the next tile is uncleared, move the head of the line forward.
+        if (get_line_orientation() == ORIENTATION_HORIZ) {
+          get_temp_ptr(struct Line)->negative_sprite_x -= 8;
+        } else {
+          get_temp_ptr(struct Line)->negative_sprite_y -= 8;
+        }
       }
     }
 
     // Now do the positive direction.
-    if (!get_line_is_positive_complete_flag_from_byte(temp_byte_6)) {
-      // Calculate positive-direction line segment origin based
-      // on negative-direction line segment origin.
-      set_positive_line_segment_origin(get_negative_line_segment_origin() +
-                                       get_tile_index_delta());
-
+    if (!get_line_is_positive_complete_flag_from_byte(get_flags_byte())) {
       // Before moving the line segment forward, update the metadata for the
       // tile we're moving from.
-      set_current_playfield_index(get_positive_line_segment_origin() +
-                                  get_temp_ptr(struct Line)->tile_step_count *
-                                      get_tile_index_delta());
+      set_current_playfield_index(playfield_tile_from_pixel_coords(get_temp_ptr(struct Line)->positive_sprite_x, get_temp_ptr(struct Line)->positive_sprite_y));
 
       // While it was the front of the line segment, current playfield tile was
       // being drawn as a sprite. Now that it's complete, update the playfield
       // and bg tile. Note: The origin tile already has the playfield flags (and
-      // bg tile) set. We can ignore that one (when tile_step_count == 0).
-      if (get_temp_ptr(struct Line)->tile_step_count != 0) {
+      // bg tile) set. We can ignore that one.
+      if (!get_line_is_first_step_flag_from_byte(get_flags_byte())) {
         set_playfield_tile(
             get_current_playfield_index(),
             get_playfield_tile_type_line(get_line_orientation(), line_index,
@@ -1066,6 +1063,11 @@ unsigned char update_line(unsigned char line_index) {
       // segment origin and update them to cleared.
       if (get_playfield_tile_type(get_current_playfield_index()) ==
           PLAYFIELD_WALL) {
+      // Calculate positive-direction line segment origin based
+      // on negative-direction line segment origin.
+      set_positive_line_segment_origin(get_temp_ptr(struct Line)->origin +
+                                       get_tile_index_delta());
+
         while (1) {
           // Walk back towards origin by one tile.
           set_current_playfield_index(get_current_playfield_index() -
@@ -1094,11 +1096,18 @@ unsigned char update_line(unsigned char line_index) {
         }
 
         set_was_line_segment_completed(TRUE);
+      } else {
+        // If the next tile is uncleared, move the head of the line forward.
+        if (get_line_orientation() == ORIENTATION_HORIZ) {
+          get_temp_ptr(struct Line)->positive_sprite_x += 8;
+        } else {
+          get_temp_ptr(struct Line)->positive_sprite_y += 8;
+        }
       }
     }
 
-    get_temp_ptr(struct Line)->current_block_completion = 0;
-    get_temp_ptr(struct Line)->tile_step_count++;
+    get_temp_ptr(struct Line)->current_block_completion = 0;    
+    unset_line_is_first_step_flag_in_byte(get_temp_ptr(struct Line)->flags);
   } else {
     if (get_frame_count() % 2 == 0) {
       ++get_temp_ptr(struct Line)->current_block_completion;
@@ -1149,6 +1158,10 @@ void start_line(unsigned char player_index) {
     lines[player_index].origin = get_negative_line_segment_origin();
     unset_line_is_negative_complete_flag(player_index);
 
+    // Track where the negative-direction head of line sprite should be drawn.
+    lines[player_index].negative_sprite_x = playfield_index_pixel_coord_x(get_negative_line_segment_origin());
+    lines[player_index].negative_sprite_y = playfield_index_pixel_coord_y(get_negative_line_segment_origin());
+
     // Now check to see if we can start a positive-direction line segment.
     set_tile_index_delta(compute_tile_index_delta(get_line_orientation()));
     set_positive_line_segment_origin(get_negative_line_segment_origin() +
@@ -1166,10 +1179,14 @@ void start_line(unsigned char player_index) {
           get_playfield_bg_tile_line_origin(get_line_orientation(),
                                             LINE_DIRECTION_POSITIVE));
       unset_line_is_positive_complete_flag(player_index);
+
+      // Track where the positive-direction head of line sprite should be drawn.
+      lines[player_index].positive_sprite_x = playfield_index_pixel_coord_x(get_positive_line_segment_origin());
+      lines[player_index].positive_sprite_y = playfield_index_pixel_coord_y(get_positive_line_segment_origin());
     }
 
     // Current line segment front tile is the origin tile.
-    lines[player_index].tile_step_count = 0;
+    set_line_is_first_step_flag(player_index);
     // The origin tiles start at complete.
     lines[player_index].current_block_completion = 8;
 
@@ -1185,16 +1202,10 @@ void draw_line(void) {
 
   if (get_line_is_started_flag_from_byte(get_flags_byte())) {
     set_line_orientation(get_line_orientation_flag_from_byte(get_flags_byte()));
-    set_tile_index_delta(compute_tile_index_delta(get_line_orientation()));
-    set_negative_line_segment_origin(get_temp_ptr(struct Line)->origin);
 
     if (!get_line_is_negative_complete_flag_from_byte(get_flags_byte())) {
-      set_current_playfield_index(get_negative_line_segment_origin() -
-                                  get_temp_ptr(struct Line)->tile_step_count *
-                                      get_tile_index_delta());
-
-      oam_spr(playfield_index_pixel_coord_x(get_current_playfield_index()),
-              playfield_index_pixel_coord_y(get_current_playfield_index()) - 1,
+      oam_spr(get_temp_ptr(struct Line)->negative_sprite_x,
+              get_temp_ptr(struct Line)->negative_sprite_y - 1,
               get_line_sprite_index(
                   get_line_orientation(),
                   get_temp_ptr(struct Line)->current_block_completion),
@@ -1202,13 +1213,8 @@ void draw_line(void) {
     }
 
     if (!get_line_is_positive_complete_flag_from_byte(get_flags_byte())) {
-      set_positive_line_segment_origin(get_negative_line_segment_origin() +
-                                       get_tile_index_delta());
-      set_current_playfield_index(get_positive_line_segment_origin() +
-                                  get_temp_ptr(struct Line)->tile_step_count *
-                                      get_tile_index_delta());
-      oam_spr(playfield_index_pixel_coord_x(get_current_playfield_index()),
-              playfield_index_pixel_coord_y(get_current_playfield_index()) - 1,
+      oam_spr(get_temp_ptr(struct Line)->positive_sprite_x,
+              get_temp_ptr(struct Line)->positive_sprite_y - 1,
               get_line_sprite_index(
                   get_line_orientation(),
                   get_temp_ptr(struct Line)->current_block_completion),
@@ -1261,9 +1267,7 @@ void check_ball_line_collisions(void) {
       // This is drawn as a sprite and we haven't updated the playfield
       // metadata to include the line flags for this tile so we don't need
       // to do anything to the playfield for this tile index.
-      set_current_playfield_index(get_negative_line_segment_origin() -
-                                  get_temp_ptr(struct Line)->tile_step_count *
-                                      get_tile_index_delta());
+      set_current_playfield_index(playfield_tile_from_pixel_coords(get_temp_ptr(struct Line)->negative_sprite_x, get_temp_ptr(struct Line)->negative_sprite_y));
 
       // Walk back across the line segment (to origin) and reset the playfield
       // tiles to uncleared.
@@ -1300,9 +1304,7 @@ void check_ball_line_collisions(void) {
       // This is drawn as a sprite and we haven't updated the playfield
       // metadata to include the line flags for this tile so we don't need
       // to do anything to the playfield for this tile index.
-      set_current_playfield_index(get_positive_line_segment_origin() +
-                                  get_temp_ptr(struct Line)->tile_step_count *
-                                      get_tile_index_delta());
+      set_current_playfield_index(playfield_tile_from_pixel_coords(get_temp_ptr(struct Line)->positive_sprite_x, get_temp_ptr(struct Line)->positive_sprite_y));
 
       // Walk back across the line segment (to origin) and reset the playfield
       // tiles to uncleared.