--- a/src/jeznes.h
+++ b/src/jeznes.h
@@ -11,6 +11,18 @@
 
 #define rand2() (rand8() % 2)
 
+// Table of tile index deltas indexed by orientation flag.
+const unsigned char tile_index_delta_table[] = {
+  1,
+  32,
+};
+
+// The tile index delta is the number we add to a playfield tile index to move
+// to the next playfield tile along the line segment.
+// |orientation| is the line orientation and must be one of ORIENTATION_HORIZ
+// or ORIENTATION_VERT.
+#define compute_tile_index_delta(orientation) (tile_index_delta_table[orientation])
+
 int main(void);
 
 // Initializes the locations and directions of the active balls on the valid