Replaces some computations with lookup tables, keeps track of the line segment head sprite position via x,y bytes instead of needing to calculate them in several places, simplifies a lot of helpers to reduce their logical workload.

Improves most of the line functions (probably others, too):
- `handle_gameplay_logic_player_and_line()`: 6016 -> 3379 cycles (~44%)
- `draw_line()`: 1976 -> 651 cycles (~67%)
- `update_line()`: 3094 -> 1793 cycles (~42%)